move tag and toggle to dedicated directories
