+++ pages/integrations/export-crm.js
@@ -6,7 +6,7 @@ import Card from "@components/Card/Card"
 import Destination from "@components/Destination"
 import Header from "@components/Structural/Header/Header"
 import { SyncStatus } from "@components/SyncStatus"
-import Toggle from "@components/Toggle"
+import Toggle from "@components/Toggle/Toggle"
 
 const accountDisplayMappings = ["Retailer type", "Annual revenue", "Local YoY growth"]
 const contactDisplayMappings = ["Job title", "Years of experience", "Personal preferences"]
